Se agrega modificaciones a los CRUD y validaciones

+++ src/main/java/com/classmanagement/controlcourses/rest/CourseRest.java
@@ -1,11 +1,16 @@
 package com.classmanagement.controlcourses.rest;
 
 import java.util.List;
+import java.util.Optional;
+
+import javax.validation.Valid;
 
 import com.classmanagement.controlcourses.dao.CourseDAO;
 import com.classmanagement.controlcourses.model.Course;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -19,21 +24,41 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("courses")
 public class CourseRest {
 
-    @Autowired
+    @Autowired //injected CourseDAO dependency
     private CourseDAO courseDAO;
 
-    @PostMapping("/guardar")
-    public void guardar(@RequestBody Course course){
-        courseDAO.save(course);
+
+    @PostMapping("/post")
+    public ResponseEntity<Course> save(@Valid @RequestBody Course course) {
+        Course obj = courseDAO.save(course);
+        return new ResponseEntity<Course>(obj, HttpStatus.OK);
     }
-    @GetMapping("/get")
+
+    @GetMapping("/get")//It's a get metod, with this you can get all Courses.
     public List<Course> get(){
         return courseDAO.findAll();
     }
-    @DeleteMapping("/delete/{id}")
-    public void delete(@PathVariable("id") Integer id){
-        courseDAO.deleteById(id);
+
+    @GetMapping("/get/{id}")//It's a get metod, with this you can get all Courses.
+    public Course retrieveCourse(@PathVariable Integer id){
+        courseDAO.findById(id);
+        return courseDAO.getOne(id);
     }
+
+    
+    @DeleteMapping("/delete/{id}")
+    public ResponseEntity<Course> delete(@PathVariable Integer id){
+        Optional<Course> cours = courseDAO.findById(id);
+        Course course = courseDAO.getOne(id);
+        courseDAO.deleteById(id);
+        if(cours.isPresent()){
+            courseDAO.deleteById(id);
+            return new ResponseEntity<Course>(course,HttpStatus.OK);
+        }else{
+            return new ResponseEntity<Course>(course,HttpStatus.NOT_FOUND);
+        }
+    }
+
     @PutMapping("/put")
     public void put(@RequestBody Course course){
         courseDAO.save(course);
